[N/A] Split app directory tests
